Base Game Finished
Completed the turtle, completed the shop menu with the dog, turtle, and charge rifle for sale (still has price of 1 goo for testing). Added 4 inventory slots that show the item when purchased in the slot and also become highlighted when selecting. Finally, added a slider for the charge rifle timer.